+++ examples/examples/subscriber-filter.rs
@@ -3,10 +3,10 @@
 mod yak_shave;
 
 fn main() {
-    use tracing_subscriber::{fmt, Filter};
+    use tracing_subscriber::{fmt, EnvFilter};
 
     let subscriber = fmt::Subscriber::builder()
-        .with_filter(Filter::from_default_env())
+        .with_env_filter(EnvFilter::from_default_env())
         .finish();
 
     tracing::subscriber::with_default(subscriber, || {