+++ examples/examples/tower-h2-client.rs
@@ -18,10 +18,14 @@ use tracing_tower::InstrumentableService;
 pub struct Conn(SocketAddr);
 
 fn main() {
+    use tracing_subscriber::filter::EnvFilter;
     // Set the default subscriber to record all traces emitted by this example
     // and by the `tracing_tower` library's helpers.
+    let filter = EnvFilter::from_default_env()
+        .add_directive("tower_h2_client=trace".parse().unwrap())
+        .add_directive("tracing_tower=trace".parse().unwrap());
     let subscriber = tracing_subscriber::FmtSubscriber::builder()
-        .with_filter("tower_h2_client=trace,tracing_tower=trace")
+        .with_env_filter(filter)
         .finish();
     let _ = tracing::subscriber::set_global_default(subscriber);